refactor: 도메인 패키지 추가
변경 내용:
- Lotto의 패키지를 변경
